borland compile
[git-p4: depot-paths = "//depot/qt/main/": change = 168048]

+++ src/gui/painting/qbrush.cpp
@@ -65,7 +65,7 @@ const uchar *qt_patternForBrush(int brushStyle, bool invert)
 QPixmap qt_pixmapForBrush(int brushStyle, bool invert)
 {
     QPixmap pm;
-    QString key = "$qt-brush$" + QString::number(brushStyle) + QString::number(invert);
+    QString key = "$qt-brush$" + QString::number(brushStyle) + QString::number((int)invert);
     if (!QPixmapCache::find(key, pm)) {
         pm = QBitmap(8, 8, qt_patternForBrush(brushStyle, invert), true);
         QPixmapCache::insert(key, pm);
